+++ CHANGELOG.md
@@ -48,6 +48,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 1.0.0
 
 -   Removed Commands: Open JAK's Website, Open JAK's API and Open JAK's Discord Bot Invite
--   Added \*.jonak Files as a Language
--   Added Syntax Higlighting for \*.jonak Files
--   Added Snippets for \*.jonak Files
+-   Added support for \*.jonak Files